AI-Powered Documentation/requirements.txt
beautifulsoup4
requests
google-generativeai
textstat
